refactor: simplify config a bit more
